modularize fn
